+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+See [ROADMAP](ROADMAP.md).
+
+## v0.6.0 - 2016-07-01
+
 ### Added
 - DocumentEventBuilder
 - DocumentTestCase assertions: assertElementHasClass, assertElementNotHasClass, assertElementsCount
@@ -16,13 +20,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - static DocumentBuilder.createFromUrl -> public DocumentBuilder.setContentFromUrl()
-- Use native method querySelectors instead jQuery.
+- Use native querySelectors method instead jQuery. 
 - Type hinting in DOM assertions based on selectors: accept NodeSelector.
 
 ### Fixed
 - Missed msg argument in DocumentTestCase assertions.
-
-See [ROADMAP](ROADMAP.md).
 
 ## v0.5.0 - 2016-06-30
 ### Added